refactor: move Board component to components/Board.tsx
